Add a getImageInVisual withSource function in the manager.
Due to the particular path of the ressource to fetch (Image in Visual),
and since the SDK architecture don't allow us to do it,
We create a new client function to get ressource on specific path.
We can now easily and efficiently fetch a 'ref' image in a specific visual.